+++ src/fileFrom/fileFrom.ts
@@ -1,8 +1,10 @@
 import {FileData, Settings} from '../types'
 import {Url} from '../api/fromUrl'
+import {Uuid} from '../api/info'
 import {UploadFromObject} from './UploadFromObject'
 import {UploadFromUrl} from './UploadFromUrl'
 import {UploadFromInterface} from './UploadFrom'
+import {UploadFromUploaded} from './UploadFromUploaded'
 
 export enum FileFrom {
   Object = 'object',
@@ -20,12 +22,14 @@ export enum FileFrom {
  * @throws Error
  * @returns {UploadFromInterface}
  */
-export default function fileFrom(from: FileFrom, data: FileData | Url, settings: Settings = {}): UploadFromInterface {
+export default function fileFrom(from: FileFrom, data: FileData | Url | Uuid, settings: Settings = {}): UploadFromInterface {
   switch (from) {
     case FileFrom.Object:
       return new UploadFromObject(data as FileData, settings)
     case FileFrom.URL:
       return new UploadFromUrl(data as Url, settings)
+    case FileFrom.Uploaded:
+      return new UploadFromUploaded(data as Uuid, settings)
     default:
       throw new Error(`File uploading from "${from}" is not supported`)
   }